Primitive Combat System
Introduced a primitive and simple combat system:
in baseEnemy Script added an attack method , that attack whenever player position and enemy position delta is lower than 1f.  There is a "attack flow control" that is almost the same of the player combat script one.
Added an attack point to enemy that has to be adjusted according to enemy attack animation.

Added the take damage method in player , that triggers the "take hit" animation